requirements.txt
Flask
Flask-RESTful
Flask-JWT-Extended
Flask-Mail
Flask-CORS
Flask-SQLAlchemy
psycopg2-binary
psycopg
# psycopg2-binary is required for Windows if you encounter DLL issues with psycopg2.
python-dotenv
bcrypt
Werkzeug
